% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{create_inline_stylesheet}
\alias{create_inline_stylesheet}
\title{Create an inline stylesheet}
\usage{
create_inline_stylesheet(css)
}
\arguments{
\item{css}{A length-one character string containing CSS.}
}
\value{
An HTML \verb{<style>} element containing the CSS.
}
\description{
Create an inline stylesheet
}
\keyword{internal}
