% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{create_remote_stylesheet_link}
\alias{create_remote_stylesheet_link}
\title{Create a remote stylesheet link}
\usage{
create_remote_stylesheet_link(remote_css)
}
\arguments{
\item{remote_css}{URL of the remote stylesheet.}
}
\value{
An HTML \verb{<link>} element.
}
\description{
Create a remote stylesheet link
}
\keyword{internal}
