% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{escape_html_attribute}
\alias{escape_html_attribute}
\title{Escape text for use in a double-quoted HTML attribute}
\usage{
escape_html_attribute(x)
}
\arguments{
\item{x}{A length-one character string.}
}
\value{
HTML-escaped text.
}
\description{
Escape text for use in a double-quoted HTML attribute
}
\keyword{internal}
