% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape text for literal use in a regular expression}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{A length-one character string.}
}
\value{
The escaped character string.
}
\description{
Escape text for literal use in a regular expression
}
\keyword{internal}
