% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{normalise_css}
\alias{normalise_css}
\title{Normalise CSS content before comparison}
\usage{
normalise_css(css)
}
\arguments{
\item{css}{A length-one character string containing CSS.}
}
\value{
A normalised length-one character string.
}
\description{
Normalise CSS content before comparison
}
\keyword{internal}
