% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{read_local_css}
\alias{read_local_css}
\title{Read a local CSS file}
\usage{
read_local_css(path)
}
\arguments{
\item{path}{Path to a local CSS file.}
}
\value{
A length-one character string containing the CSS.
}
\description{
Read a local CSS file
}
\keyword{internal}
