% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_self_contained.R
\name{remove_self_contained}
\alias{remove_self_contained}
\title{Remove self-contained option from R Markdown YAML}
\usage{
remove_self_contained(doc)
}
\arguments{
\item{doc}{A character string. Path to an input \code{.Rmd} file.
The file must contain a valid YAML header delimited by \verb{---}.}
}
\value{
Invisibly returns the path to the modified \code{.Rmd} file.
}
\description{
Reads an R Markdown (\code{.Rmd}) file, modifies the YAML header to ensure
\code{self_contained: no} is set for \code{html_document}, and writes a modified
copy to the same directory with a \verb{_tiny.Rmd} suffix.
}
\details{
If \code{self_contained} is not present in the YAML, it will be added.
If it is present and set to \code{yes}, it will be changed to \code{no}.
}
\examples{
\dontrun{
remove_self_contained("report.Rmd")
}

}
