% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tiny_script.R
\name{remove_tiny_script}
\alias{remove_tiny_script}
\title{Remove generated \verb{_tiny.Rmd} file}
\usage{
remove_tiny_script(doc)
}
\arguments{
\item{doc}{A character string. Path to an input \code{.Rmd} file.
The corresponding \verb{_tiny.Rmd} file is assumed to have the same name
with a \verb{_tiny} suffix in the same directory.}
}
\value{
Invisibly returns \code{TRUE} if the file was deleted, \code{FALSE} if the file
did not exist or could not be removed.
}
\description{
Deletes the \verb{_tiny.Rmd} file associated with a given R Markdown (\code{.Rmd})
input, if it exists in the same directory.
}
\examples{
\dontrun{
remove_tiny_script("report.Rmd")
}

}
