% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{replace_matching_links}
\alias{replace_matching_links}
\title{Replace matching stylesheet links}
\usage{
replace_matching_links(html, pattern, replacement)
}
\arguments{
\item{html}{Character vector containing HTML.}

\item{pattern}{Regular expression matching stylesheet links.}

\item{replacement}{Replacement HTML.}
}
\value{
The updated HTML character vector.
}
\description{
Replace matching stylesheet links
}
\keyword{internal}
