% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{stylesheet_link_pattern}
\alias{stylesheet_link_pattern}
\title{Construct a stylesheet link-matching regular expression}
\usage{
stylesheet_link_pattern(linked_css)
}
\arguments{
\item{linked_css}{CSS filename or relative path to locate.}
}
\value{
A Perl-compatible regular expression.
}
\description{
Construct a stylesheet link-matching regular expression
}
\keyword{internal}
