% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_knit_addin.R
\name{tiny_knit_addin}
\alias{tiny_knit_addin}
\title{Run tiny_knit on the currently open R Markdown file}
\usage{
tiny_knit_addin()
}
\description{
This addin runs \code{\link[=tiny_knit]{tiny_knit()}} on the R Markdown file open in the
RStudio Source pane.
}
