% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_stylesheet.R
\name{validate_stylesheet_inputs}
\alias{validate_stylesheet_inputs}
\title{Validate stylesheet replacement inputs}
\usage{
validate_stylesheet_inputs(
  html,
  local_css,
  remote_css,
  linked_css,
  timeout,
  warn
)
}
\arguments{
\item{html}{Character vector containing HTML.}

\item{local_css}{Path to the local CSS file.}

\item{remote_css}{Raw URL of the remote CSS file.}

\item{linked_css}{CSS filename or relative path to locate in the HTML.}

\item{timeout}{Connection timeout in seconds.}

\item{warn}{Logical controlling warning messages.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate stylesheet replacement inputs
}
\keyword{internal}
